Patent analysis the usage of the Google Patents Public Datasets on BigQuery
The contents of this repository aren't an legit Google product.
 
Google Patents Public Datasets is a group of well suited BigQuery database tables from government, studies and non-public organizations for engaging in statistical analysis of patent statistics. The records is available to be queried with SQL thru BigQuery, joined with non-public datasets you upload, and exported and processed the usage of many other like minded analysis tools. This repository is a centralized source for examples which use the information.
 
Currently the repo consists of  examples:
Patent Landscaping: A demo of an automated method of locating patents associated to a selected topic given an preliminary seed set of patents. Based on the paper by way of Dave Feltenberger and Aaron Abood, Automated Patent Landscaping.
Claim Text Extraction: A demo of interacting with patent claim text facts using BigQuery and python.
 
Claim Breadth Model: A system learning method for estimating patent claim breadth the usage of records from BigQuery.